update: NHF schema and ID types for NHF v1.2.0 (#95)

+++ src/icefabric/hydrofabric/subset_nhf.py
@@ -31,9 +31,9 @@ def _build_upstream_dict_from_nexus(
     """Build upstream connectivity dictionary from flowpath nexus connections."""
     fp_pl = flowpaths_pl.with_columns(
         [
-            pl.col(edge_id).cast(pl.Int32),
-            pl.col(f"up_{node_id}").cast(pl.Int32),
-            pl.col(f"dn_{node_id}").cast(pl.Int32),
+            pl.col(edge_id).cast(pl.Int64),
+            pl.col(f"up_{node_id}").cast(pl.Int64),
+            pl.col(f"dn_{node_id}").cast(pl.Int64),
         ]
     )
     nexus_to_downstream = fp_pl.select(
